correccion fit abs()
